requirements.txt
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.2.0
xgboost>=1.7.0
pyyaml>=6.0
pytest>=7.0
joblib>=1.2.0
